Fix Windows build: Change generation type from int to size_t
- Fixed MSVC error C2220 (warning C4267 treated as error)
- Changed updateStatistics parameter from int to size_t
- Changed Statistics::generation field from int to size_t
- Resolves type conversion warning on Windows builds
- Unblocks Windows packaging workflows (Chocolatey, NSIS)

This fix ensures consistent type usage throughout the evolutionary
optimizer and eliminates implicit size_t to int conversions that
MSVC flags as potential data loss.

+++ cpp/evolutionary/src/evolutionary.cpp
@@ -791,7 +791,7 @@ bool EvolutionaryOptimizer::checkStagnation() const {
     return (recentBest - oldBest) < config_.stagnationThreshold;
 }
 
-void EvolutionaryOptimizer::updateStatistics(int generation) {
+void EvolutionaryOptimizer::updateStatistics(size_t generation) {
     Statistics stats;
     stats.generation = generation;
     stats.bestFitness = population_.getBestFitness();
